sched: add a new arch_sd_local_flags for sched_domain init
Date	Fri, 18 Oct 2013 13:52:15 +0200

The function arch_sd_local_flags is used to set flags in sched_domains
according to the platform architecture. A new flag SD_SHARE_POWERDOMAIN
is
also created to reflect whether groups of CPUs in a sched_domain level
can or
not reach different power state. As an example, the flag should be
cleared at
CPU level if groups of cores can be power gated independently. This
information
is used to decide if it's worth packing some tasks in a group of CPUs
in order
to power gate the other groups instead of spreading the tasks. The
default
behavior of the scheduler is to spread tasks across CPUs and groups of
CPUs so
the flag is set into all sched_domains.

The cpu parameter of arch_sd_local_flags can be used by architecture to
fine
tune the scheduler domain flags. As an example SD_SHARE_POWERDOMAIN
flag can be
set differently for groups of CPUs according to DT information

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>

Signed-off-by: Paul Reioux <reioux@gmail.com>

Conflicts:
	include/linux/topology.h